backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
PyMySQL==1.1.1
pydantic==2.11.4
pydantic-settings==2.9.1
PyJWT==2.9.0
cryptography==44.0.3
